track_sys/asset-tracker-backend/requirements.txt
alembic==1.15.2
aniso8601==10.0.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
# Flask==3.1.0
# flask-cors==5.0.1
# Flask-JWT-Extended==4.7.1
# Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
# Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
packaging==25.0
pluggy==1.5.0
PyJWT==2.10.1
pytest==8.3.5
# python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
# SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
Flask-CORS==3.0.10
python-dotenv==0.19.0
SQLAlchemy==1.4.23